Update build tags for more platforms
