+++ galois_arm64.go
@@ -1,18 +1,17 @@
-//go:build !noasm && !appengine && !gccgo && pshufb
+//go:build !noasm && !appengine && !gccgo && !nopshufb
 
 // Copyright 2015, Klaus Post, see LICENSE for details.
 // Copyright 2017, Minio, Inc.
 
 package reedsolomon
 
+const pshufb = true
+
 //go:noescape
 func galMulNEON(low, high, in, out []byte)
 
 //go:noescape
 func galMulXorNEON(low, high, in, out []byte)
-
-//go:noescape
-func galXorNEON(in, out []byte)
 
 func galMulSlice(c byte, in, out []byte, o *options) {
 	if c == 1 {
@@ -50,20 +49,6 @@ func galMulSliceXor(c byte, in, out []byte, o *options) {
 	}
 }
 
-// simple slice xor
-func sliceXor(in, out []byte, o *options) {
-
-	galXorNEON(in, out)
-	done := (len(in) >> 5) << 5
-
-	remain := len(in) - done
-	if remain > 0 {
-		for i := done; i < len(in); i++ {
-			out[i] ^= in[i]
-		}
-	}
-}
-
 // 4-way butterfly
 func ifftDIT4(work [][]byte, dist int, log_m01, log_m23, log_m02 ffe, o *options) {
 	ifftDIT4Ref(work, dist, log_m01, log_m23, log_m02, o)
@@ -89,7 +74,7 @@ func fftDIT2(x, y []byte, log_m ffe, o *options) {
 	// Reference version:
 	refMulAdd(x, y, log_m)
 	// 64 byte aligned, always full.
-	galXorNEON(x, y)
+	xorSliceNEON(x, y)
 }
 
 // 2-way butterfly forward
@@ -102,7 +87,7 @@ func fftDIT28(x, y []byte, log_m ffe8, o *options) {
 // 2-way butterfly
 func ifftDIT2(x, y []byte, log_m ffe, o *options) {
 	// 64 byte aligned, always full.
-	galXorNEON(x, y)
+	xorSliceNEON(x, y)
 	// Reference version:
 	refMulAdd(x, y, log_m)
 }
