+++ galois_ppc64le.go
@@ -1,9 +1,11 @@
-//go:build !noasm && !appengine && !gccgo && !pshufb
+//go:build !noasm && !appengine && !gccgo && !nopshufb
 
 // Copyright 2015, Klaus Post, see LICENSE for details.
 // Copyright 2018, Minio, Inc.
 
 package reedsolomon
+
+const pshufb = true
 
 //go:noescape
 func galMulPpc(low, high, in, out []byte)
