+++ xor_noasm.go
@@ -1,0 +1,7 @@
+//go:build noasm || gccgo || appengine || (ppc64le && nopshufb)
+
+package reedsolomon
+
+func sliceXor(in, out []byte, o *options) {
+	sliceXorGo(in, out, o)
+}
